// src/app/common/ui-component.interface.ts
export interface IUIComponentTemplate{
    textBox:string;
    checkBox:boolean;
    dropDown:any;
    LabelName:string;
    istextbox: boolean;
    isdropDown: boolean;
    DropdownArray: string[];
    textType: string;
    ismandatory: boolean;
    OPTM_SEQ: number;
    AliasID: string;
    Size: number
}